Implement removal of multiparm instances.
This is done very similarly to addition.

Also, change HAPI_Parm::getMultiparmTemplate to HAPI_Parm::getTemplate
to properly reflect its semantics.

Additionally, generate the control for the multiparm before getting its
value from the host, since it might have changed. Ultimately, we need to
cache the integer parameter without setting it into the host until the
user has finished changing the parameter, otherwise multiparms will be
destroyed as the user types in the integer. This will require a massive
restructuring of the code in HAPI_AssetGUI.cs, which is to come much later.