Added HAPI_CookOptions to control whether to refine curves inside HAPI or return just the control points and have the client refine the curves using the new HAPI_CurveInfo structs. We decided to include the maximum vertices per primitive setting in the cook options and the curve refinement LOD. We removed the minimum vertices per primitive setting altogether because it was not actually being enforced.
The API changed by requiring the need for a HAPI_CookOptions input to HAPI_Initialize to be used as the global cook options and to HAPI_CookAsset for individual cooks (passing NULL here will use the global options). Removed min/max vertices per primitive parameters from HAPI_InstantiateAsset and HAPI_LoadHIPFile. Also added the cook_on_load parameter to HAPI_LoadHIPFile.

Another side-effect of this change is that it is now possible to avoid convexing output geometry entirely. We use convexing to restrict the maximum number of vertices per primitive. This operation was always being applied. But now, if you do not have a maximum to worry about than you can just pass in -1 as the maximum vertices per primitive and the convexing step will be omitted, giving a slight performance improvement.