Removed HAPI_AssetInfo::minTransInputCount and HAPI_AssetInfo::minGeoInputCount. As a result, renamed HAPI_AssetInfo::maxTransInputCount to transformInputCount and HAPI_AssetInfo::maxGeoInputCount to geoInputCount.
We did this because, now that we count certain node path parameters as inputs in addition to the actual node inputs, we need to know WHICH inputs are required. It is no longer enough to just have a minimum integer value. There is no way to tag inputs as required right now but the min input counts no longer made sense either way.